requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
gspread==6.0.2
oauth2client==4.1.3
fastapi
uvicorn